requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
line-bot-sdk>=3.1.0